+++ src/constants/sectors.ts
@@ -13,16 +13,18 @@
  */
 export const DataspaceSectors = Object.freeze({
   HealthCare: 'health-care',
+  HealthLab: 'health-lab',
   HealthResearch: 'health-research',
   HealthTech: 'health-tech',
   HealthInsurance: 'health-insurance',
   AnimalCare: 'animal-care',
+  AnimalLab: 'animal-lab',
   AnimalResearch: 'animal-research',
   AnimalInsurance: 'animal-insurance',
   AnimalTech: 'animal-tech',
   OneHealthResearch: 'onehealth-research',
+  OneHealthLab: 'onehealth-lab',
   OneHealthTech: 'onehealth-tech',
 } as const);
 
 export type DataspaceSector = typeof DataspaceSectors[keyof typeof DataspaceSectors];
-